Extract defra-node benchmark tests
